Move contracts code to contract.rs
