+++ crates/scarb-api/src/lib.rs
@@ -1,84 +1,23 @@
-use anyhow::{anyhow, Context, Result};
+use anyhow::{anyhow, Result};
 use camino::{Utf8Path, Utf8PathBuf};
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 use scarb_metadata::{Metadata, PackageId, PackageMetadata, TargetMetadata};
 use semver::VersionReq;
-use serde::Deserialize;
 use std::collections::HashMap;
-use std::fs;
 use std::str::FromStr;
-use universal_sierra_compiler_api::{compile_sierra_at_path, SierraType};
 
+use crate::contracts::{
+    load_contracts_artifacts_and_source_sierra_paths, StarknetContractArtifacts,
+};
 pub use command::*;
 use shared::print::print_as_warning;
 
 mod command;
+mod contracts;
 pub mod metadata;
 pub mod version;
 
 const INTEGRATION_TEST_TYPE: &str = "integration";
-
-#[derive(Deserialize, Debug, PartialEq, Clone)]
-struct StarknetArtifacts {
-    version: u32,
-    contracts: Vec<StarknetContract>,
-}
-
-#[allow(dead_code)]
-#[derive(Deserialize, Debug, PartialEq, Clone)]
-struct StarknetContract {
-    id: String,
-    package_name: String,
-    contract_name: String,
-    artifacts: StarknetContractArtifactPaths,
-}
-
-#[allow(dead_code)]
-#[derive(Deserialize, Debug, PartialEq, Clone)]
-struct StarknetContractArtifactPaths {
-    sierra: Utf8PathBuf,
-}
-
-/// Contains compiled Starknet artifacts
-#[derive(Debug, PartialEq, Clone)]
-pub struct StarknetContractArtifacts {
-    /// Compiled sierra code
-    pub sierra: String,
-    /// Compiled casm code
-    pub casm: String,
-}
-
-impl StarknetContractArtifacts {
-    fn from_scarb_contract_artifact(
-        starknet_contract: &StarknetContract,
-        base_path: &Utf8Path,
-    ) -> Result<Self> {
-        let sierra_path = base_path.join(starknet_contract.artifacts.sierra.clone());
-        let sierra = fs::read_to_string(sierra_path)?;
-
-        let casm = compile_sierra_at_path(
-            starknet_contract.artifacts.sierra.as_str(),
-            Some(base_path.as_std_path()),
-            &SierraType::Contract,
-        )?;
-
-        Ok(Self { sierra, casm })
-    }
-}
-
-/// Get deserialized contents of `starknet_artifacts.json` file generated by Scarb
-///
-/// # Arguments
-///
-/// * `path` - A path to `starknet_artifacts.json` file.
-fn artifacts_for_package(path: &Utf8Path) -> Result<StarknetArtifacts> {
-    let starknet_artifacts =
-        fs::read_to_string(path).with_context(|| format!("Failed to read {path:?} contents"))?;
-    let starknet_artifacts: StarknetArtifacts =
-        serde_json::from_str(starknet_artifacts.as_str())
-            .with_context(|| format!("Failed to parse {path:?} contents. Make sure you have enabled sierra code generation in Scarb.toml"))?;
-    Ok(starknet_artifacts)
-}
 
 #[derive(PartialEq, Debug)]
 struct StarknetArtifactsFiles {
@@ -219,27 +158,6 @@ pub fn get_contracts_artifacts_and_source_sierra_paths(
     }
 }
 
-fn load_contracts_artifacts_and_source_sierra_paths(
-    contracts_path: &Utf8PathBuf,
-) -> Result<HashMap<String, (StarknetContractArtifacts, Utf8PathBuf)>> {
-    let base_path = contracts_path
-        .parent()
-        .ok_or_else(|| anyhow!("Failed to get parent for path = {}", &contracts_path))?;
-    let artifacts = artifacts_for_package(contracts_path)?;
-    let mut map = HashMap::new();
-
-    for ref contract in artifacts.contracts {
-        let name = contract.contract_name.clone();
-        let contract_artifacts =
-            StarknetContractArtifacts::from_scarb_contract_artifact(contract, base_path)?;
-
-        let sierra_path = base_path.join(contract.artifacts.sierra.clone());
-
-        map.insert(name.clone(), (contract_artifacts, sierra_path));
-    }
-    Ok(map)
-}
-
 #[must_use]
 pub fn target_dir_for_workspace(metadata: &Metadata) -> Utf8PathBuf {
     metadata
@@ -302,13 +220,13 @@ mod tests {
     use super::*;
     use crate::metadata::MetadataCommandExt;
     use assert_fs::fixture::{FileWriteStr, PathChild, PathCopy};
-    use assert_fs::prelude::FileTouch;
     use assert_fs::TempDir;
     use camino::Utf8PathBuf;
     use indoc::{formatdoc, indoc};
+    use std::fs;
     use std::str::FromStr;
 
-    fn setup_package(package_name: &str) -> TempDir {
+    pub(crate) fn setup_package(package_name: &str) -> TempDir {
         let temp = TempDir::new().unwrap();
         temp.copy_from(
             format!("tests/data/{package_name}"),
@@ -633,41 +551,6 @@ mod tests {
             "basic_package",
         );
         assert!(path.is_none());
-    }
-
-    #[test]
-    fn parsing_starknet_artifacts() {
-        let temp = setup_package("basic_package");
-
-        ScarbCommand::new_with_stdio()
-            .current_dir(temp.path())
-            .arg("build")
-            .run()
-            .unwrap();
-
-        let artifacts_path = temp
-            .path()
-            .join("target/dev/basic_package.starknet_artifacts.json");
-        let artifacts_path = Utf8PathBuf::from_path_buf(artifacts_path).unwrap();
-
-        let artifacts = artifacts_for_package(&artifacts_path).unwrap();
-
-        assert!(!artifacts.contracts.is_empty());
-    }
-
-    #[test]
-    fn parsing_starknet_artifacts_on_invalid_file() {
-        let temp = TempDir::new().unwrap();
-        temp.copy_from("../../", &[".tool-versions"]).unwrap();
-        let path = temp.child("wrong.json");
-        path.touch().unwrap();
-        path.write_str("\"aa\": {}").unwrap();
-        let artifacts_path = Utf8PathBuf::from_path_buf(path.to_path_buf()).unwrap();
-
-        let result = artifacts_for_package(&artifacts_path);
-        let err = result.unwrap_err();
-
-        assert!(err.to_string().contains(&format!("Failed to parse {artifacts_path:?} contents. Make sure you have enabled sierra code generation in Scarb.toml")));
     }
 
     #[test]
